slight include cleanup
