+++ explore/flip_social_explore.h
@@ -1,7 +1,6 @@
 #ifndef FLIP_SOCIAL_EXPLORE_H
 #define FLIP_SOCIAL_EXPLORE_H
 #include "flip_social.h"
-#include <callback/callback.h>
 bool flip_social_get_explore(FlipperHTTP *fhttp);
 bool flip_social_get_explore_2(FlipperHTTP *fhttp);
 bool flip_social_parse_json_explore(FlipperHTTP *fhttp);
