+++ feed/flip_social_feed.h
@@ -1,8 +1,6 @@
 #ifndef FLIP_SOCIAL_FEED_H
 #define FLIP_SOCIAL_FEED_H
 #include "flip_social.h"
-#include <callback/callback.h>
-#include <flip_storage/flip_social_storage.h>
 
 bool flip_social_get_feed(FlipperHTTP *fhttp, int series_index);
 FlipSocialFeedMini *flip_social_parse_json_feed(FlipperHTTP *fhttp);
