+++ flip_social.h
@@ -327,4 +327,6 @@ extern uint8_t flip_social_feed_type_index;
 //
 extern char *flip_social_notification_type[];
 extern uint8_t flip_social_notification_type_index;
+//
+extern bool went_to_friends;
 #endif
